Build DLL时判断路径

+++ Unity/Assets/Scripts/Editor/BuildEditor/BuildAssembliesHelper.cs
@@ -15,16 +15,22 @@ namespace ET
 
         public static void BuildModel(CodeOptimization codeOptimization, GlobalConfig globalConfig)
         {
-            string[] logicFiles = Directory.GetFiles(Define.BuildOutputDir, "Model_*");
-            foreach (string file in logicFiles)
+            if (Directory.Exists(Define.BuildOutputDir))
             {
-                File.Delete(file);
+                string[] logicFiles = Directory.GetFiles(Define.BuildOutputDir, "Model_*");
+                foreach (string file in logicFiles)
+                {
+                    File.Delete(file);
+                }
             }
-            
-            logicFiles = Directory.GetFiles(CodeDir, "Model_*");
-            foreach (string modelFile in logicFiles)
+
+            if (Directory.Exists(CodeDir))
             {
-                File.Delete(modelFile);
+                string[] logicFiles = Directory.GetFiles(CodeDir, "Model_*");
+                foreach (string modelFile in logicFiles)
+                {
+                    File.Delete(modelFile);
+                } 
             }
             
             List<string> codes;
@@ -75,17 +81,24 @@ namespace ET
 
         public static void BuildHotfix(CodeOptimization codeOptimization, GlobalConfig globalConfig)
         {
-            string[] logicFiles = Directory.GetFiles(Define.BuildOutputDir, "Hotfix_*");
-            foreach (string file in logicFiles)
+            if (Directory.Exists(Define.BuildOutputDir))
             {
-                File.Delete(file);
+                string[] logicFiles = Directory.GetFiles(Define.BuildOutputDir, "Hotfix_*");
+                foreach (string file in logicFiles)
+                {
+                    File.Delete(file);
+                }
             }
-            
-            logicFiles = Directory.GetFiles(CodeDir, "Hotfix_*");
-            foreach (string modelFile in logicFiles)
+
+            if (Directory.Exists(CodeDir))
             {
-                File.Delete(modelFile);
+                string[] logicFiles = Directory.GetFiles(CodeDir, "Hotfix_*");
+                foreach (string modelFile in logicFiles)
+                {
+                    File.Delete(modelFile);
+                }
             }
+           
 
             List<string> codes;
             switch (globalConfig.CodeMode)
@@ -121,7 +134,7 @@ namespace ET
             EditorUtility.SetDirty(globalConfig);
             AssetDatabase.SaveAssets();
 
-            BuildAssembliesHelper.BuildMuteAssembly(assemblyName, codes, new[] { Path.Combine(Define.BuildOutputDir, $"Model_{globalConfig.ModelVersion}.dll") }, codeOptimization,
+            BuildAssembliesHelper.BuildMuteAssembly(assemblyName, codes, new[] { Path.Combine(CodeDir, $"Model_{globalConfig.ModelVersion}.dll.bytes") }, codeOptimization,
                 globalConfig.CodeMode);
 
             File.Copy(Path.Combine(Define.BuildOutputDir, $"{assemblyName}.dll"), Path.Combine(CodeDir, $"{assemblyName}.dll.bytes"), true);
